requirements.txt
# Core ML
torch==2.9.1
torchvision==0.24.1
transformers==4.57.3
datasets==4.4.2
evaluate==0.4.6
accelerate==1.12.0
huggingface-hub==0.36.0
tokenizers==0.22.2


# OCR
pytesseract==0.3.13
Pillow==12.1.0
opencv-python==4.12.0.88


# Data processing
numpy==2.2.6
pandas==2.3.3
scikit-learn==1.8.0
RapidFuzz==3.14.3


# Evaluation
seqeval==1.2.2


# Visualization
matplotlib==3.10.8
seaborn>=0.12.0  


# Utilities
tqdm==4.67.1
pyyaml==6.0.3
regex==2025.11.3
requests==2.32.5


# Development / optional
jupyter-client==8.7.0
ipykernel==7.1.0
black>=23.0.0
pytest>=7.4.0